# Speedup function analysis