https://github.com/radare/radare2/issues/5274#issuecomment-230157691

Started:
r2 -S ../bins/torivahti.exe -qc 'aac;Ps b'  197.53s user 0.16s system 98% cpu 3:31.82 total

Then:
<del>r2 -S ../bins/torivahti.exe -qc 'aac;Ps b'  56.20s user 0.13s system 98% cpu 57.690 total</del>
<del>r2 -S ../bins/torivahti.exe -qc 'aac;Ps b'  31.13s user 0.13s system 98% cpu 31.690 total</del>
<del>r2 -S ../bins/torivahti.exe -qc 'aac;Ps b'  25.29s user 0.13s system 98% cpu 25.758 total</del>
r2 -S ../bins/torivahti.exe -qc 'aac;Ps b'  22.97s user 0.12s system 98% cpu 23.362 total

Check results:
```
$ r2-master -S ../bins/torivahti.exe -qc 'aac;Ps a'                            
r2 -S ../bins/torivahti.exe -qc 'aac;Ps b'  197.53s user 0.16s system 98% cpu 3:31.82 total
$ r2 -S ../bins/torivahti.exe -qc 'aac;Ps b'  
r2 -S ../bins/torivahti.exe -qc 'aac;Ps a'  46.36s user 0.36s system 22% cpu 3:30.21 total
$  diff ~/.config/radare2/projects/b/rc ~/.config/radare2/projects/a/rc
< "e cfg.user = pid12045"
---
> "e cfg.user = pid843"
103561,103562c103561,103562
< f sym.___chkstk 50 0x00644a50
< af+ 0x00644a50 50 sym.___chkstk f n
---
> f sym.__alloca 50 0x00644a50
> af+ 0x00644a50 50 sym.__alloca f n

$ grep "^f" ~/.config/radare2/projects/b/rc | wc -l
29989

```

Todo:

- [ ] Fix r2r Regressions
- [x] Test if  flag_get_at can use get_i when `closest=false` instead, so all callsites enjoy (moved to #6757)