@radare 2017-02-08
get_i = gets the flag at the given address (1 hahstable lookup) (O(1))

get-at = gets closest flag at given address (O(n))

> On 08 Feb 2017, at 12:37, SchumBlubBlub <notifications@github.com> wrote:
> 
> @SchumBlubBlub commented on this pull request.
> 
> In libr/core/canal.c <https://github.com/radare/radare2/pull/6724>:
> 
> > @@ -472,7 +472,7 @@ static int core_anal_fcn(RCore *core, ut64 at, ut64 from, int reftype, int depth
>  	}
>  	fcn->addr = at;
>  	r_anal_fcn_set_size (fcn, 0);
> -	RFlagItem *fi = r_flag_get_at (core->flags, at, false);
> +	RFlagItem *fi = r_flag_get_i (core->flags, at);
> I thought that's equivalent to get_at(false)?
> 
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub <https://github.com/radare/radare2/pull/6724>, or mute the thread <https://github.com/notifications/unsubscribe-auth/AA3-lnGQFZ7DEyZ7HrYbva1EmzXWR0U-ks5raakCgaJpZM4L6LU4>.
>