@SchumBlubBlub 2017-02-08
Sounds good. Need to get rid of ALL linked lists too.