@radare 2017-02-08
yup

> On 08 Feb 2017, at 12:59, SchumBlubBlub <notifications@github.com> wrote:
> 
> Sounds good. Need to get rid of ALL linked lists too.
> 
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub <https://github.com/radare/radare2/pull/6724#issuecomment-278309155>, or mute the thread <https://github.com/notifications/unsubscribe-auth/AA3-lrSUPpDn2eS9DEb6RlOkI8ODu_DFks5raa4JgaJpZM4L6LU4>.
>