@radare 2017-02-08
with closest = false, its still iterating over all keys, ignoring the ones that are not in the current space… wondering if we can just if (!closest) return r_flag_get_i() without regressions. can u do a pr with this change to evaluate the change?


R_API RFlagItem *r_flag_get_at(RFlag *f, ut64 off, bool closest) {
        RFlagItem *item, *nice = NULL;
        RListIter *iter;

        r_list_foreach (f->flags, iter, item) {
                if (f->space_strict && IS_IN_SPACE (f, item)) {
                        continue;
                }
                if (item->offset == off) {
                        return evalFlag (f, item);
                }
                if (closest && off > item->offset) {
                        if (!nice || nice->offset < item->offset) {
                                nice = item;
                        }
                }
        }
        return evalFlag (f, nice);
}

> On 08 Feb 2017, at 13:03, SchumBlubBlub <notifications@github.com> wrote:
> 
> @SchumBlubBlub commented on this pull request.
> 
> In libr/core/canal.c <https://github.com/radare/radare2/pull/6724>:
> 
> > @@ -472,7 +472,7 @@ static int core_anal_fcn(RCore *core, ut64 at, ut64 from, int reftype, int depth
>  	}
>  	fcn->addr = at;
>  	r_anal_fcn_set_size (fcn, 0);
> -	RFlagItem *fi = r_flag_get_at (core->flags, at, false);
> +	RFlagItem *fi = r_flag_get_i (core->flags, at);
> R_API RFlagItem *r_flag_get_at(RFlag *f, ut64 off, bool closest)
> 
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub <https://github.com/radare/radare2/pull/6724>, or mute the thread <https://github.com/notifications/unsubscribe-auth/AA3-lqMHEXYbw8vjdEgJBBNiLADup-rtks5raa8jgaJpZM4L6LU4>.
>