@condret 2017-02-08
maybe we can do something similar to the pcache storage instead of the lists, or what about trees