@radare 2017-02-09
There are several regressions here, mainly for function signatures , see travis output