@Maijin 2017-02-09
It also doesn't build on OSX https://travis-ci.org/radare/radare2/jobs/199676724#L150