@SchumBlubBlub 2017-02-10
@radare, you asked me to change this to that `r_flag_get_at` uses `r_flag_get_i` for `!closest`, which is much better then changing the calls directly. But get_flag_i depends on the r_hashtable64, which is buggy so this cause regressions, and you didn't want fixes for that. So I split such part (most of the perf win here) into https://github.com/radare/radare2/issues/6757 for after flags refactor.