@radare 2017-02-10
27 code analysis regressions. we need to check them to see if the test was wrong before or its an issue in this PR. i think master will be green again tonight, so you can rebase after the rest of changes and see if travis situation improves.