@radare 2017-02-10
Optionally if you want this PR to be merged now i would suggest to add another config variable under the `anal.` eval space disabled by default. and users can activate that to test it instead of enforcing this change as default.

There are several other anal config vars that should be reviewed and activated/disabled/removed if you are in the mood. As well as improving the new `aab` analysis.

Thanks